+++ src/routes/Project/actions/TableActions.js
@@ -1,6 +1,7 @@
 import ProjectConstants from '../constants';
 
 export function addClipToTrack(clip) {
+  console.log('adding clip to track');
   return {
     type: ProjectConstants.ADD_CLIP_TO_TRACK,
     payload: clip
