+++ src/routes/Project/components/Table.js
@@ -5,13 +5,15 @@ import './Table.scss';
 export const Table = (props) => (
   <div id="table-component">
     { Object.keys(props.table).reduce((previous, currTrack) => {
+        console.log('currTrack', currTrack, props.table[currTrack]);
         previous.push(
-          <div className="track" key={currTrack}>
+          <div className="track" key={ currTrack }>
             { Object.keys(props.table[currTrack]).reduce((previous, currTime, i) => {
-                if (i % 2 !== 0) return previous;
+                //if (i % 2 !== 0) return previous;
                 previous.push(
-                  <h3>props.table[currTrack][currTime].id</h3>
-                );
+                    <h3>K:FDLK:ASLDK</h3>
+                  );
+                return previous;
               }, [])
             }
           </div>
